pyproject.toml
[tool.poetry]
name = "web-scraping-er-v2"
version = "0.1.0"
description = ""
authors = ["myuR <myurpvn@gmail.com>"]
readme = "README.md"
packages = [{include = "web_scraping_er_v2", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.1"
requests = "^2.31.0"
typer = "^0.9.0"
pyyaml = "^6.0.1"
google-cloud-bigquery = "^3.19.0"
pyarrow = "^15.0.1"
structlog = "^24.1.0"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
ipykernel = "^6.29.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
exchange-rates-pipeline = 'src.web_scraping_er_v2.__main__:app'